main.c: Send signal to Python program
